Added a somewhat crappy and broken attempt at proof.
It doesn't work because in some cases I don't have values since
I don't keep track of what the arguments that might be passed to
functions are for abstraction.  The property that I want to prove has to
be true for all possible arguments, so I don't know how to reasonably
prove that in a finite amount of space.